widen the latency accessor to the type its callers compute in
`OccupancyProblem::latencyOf` returned `unsigned`, following CIRCT's own
optional, but every caller composes it into an int64 expression or a
`LinearExpr`, and one of them subtracted. `latencyOf(op) - 1` promotes the
literal to `unsigned` and evaluates a combinational operator's zero latency as
4294967295, which is the defect `a34c059` fixed by charging the right number.
This makes the shape that produced it unrepresentable rather than absent.

`unsigned - int64_t` was already safe, since the usual arithmetic conversions
widen to int64_t; only `unsigned - <int literal>` was not, and a signed
accessor leaves neither.

The one caller that needed a change is the acyclic solve's horizon, an
`unsigned` accumulator over the same value. Nothing else moves and no value
changes: the bed is 112 of 112 byte-identical under the heuristic and 64 of 64
under the exact scheduler, `tests/` 665 passed / 42 skipped.

+++ mlir/include/allo/Scheduling/Scheduler.h
@@ -60,7 +60,11 @@ public:
   /// arrived: the latency of the operator type \p op is linked to. Every
   /// operation carries one, `populateOperatorTypes` having linked every
   /// operation the problem holds.
-  unsigned latencyOf(Operation *op);
+  ///
+  /// Signed, though the underlying latency is not: every caller composes it
+  /// into an expression that subtracts, and an `unsigned` one silently
+  /// evaluates `latencyOf(op) - 1` on a combinational operator as 2^32 - 1.
+  int64_t latencyOf(Operation *op);
 
   /// The schedule DEPTH of a SOLVED problem: the cycle by which every operation
   /// has completed. A REPORT only, since a span composes from the drain
